refactor: rename to top-moments and remove postId screen
