--- apps/mobile/src/app/(tabs)/(map)/event/post/[postId].tsx
@@ -1,1 +1,0 @@
-export { default } from '@/features/posts/post-details-screen';
